requirements.txt
aiohappyeyeballs~=2.6.1
aiohttp~=3.12.12
aiosignal~=1.3.2
annotated-types~=0.7.0
anyio~=4.9.0
attrs~=25.3.0
azure-core~=1.34.0
azure-identity~=1.15.0
blinker~=1.9.0
cachelib~=0.13.0
cachetools~=5.5.2
certifi~=2025.6.15
cffi~=1.17.1
charset-normalizer~=3.4.2
click~=8.2.1
contourpy~=1.3.2
cryptography~=45.0.4
cycler~=0.12.1
databricks~=0.2
databricks-sdk~=0.57.0
databricks-sql-connector~=4.0.5
dataclasses-json~=0.6.7
distro~=1.9.0
et_xmlfile~=2.0.0
faiss-cpu~=1.11.0.post1
fastapi~=0.116.1
filelock~=3.18.0
Flask~=3.0.0
Flask-Session~=0.8.0
fonttools~=4.58.4
frozenlist~=1.7.0
fsspec~=2025.7.0
google-auth~=2.40.3
greenlet~=3.2.3
gunicorn~=21.2.0
h11~=0.16.0
hf-xet~=1.1.5
httpcore~=1.0.9
httpx~=0.28.1
httpx-sse~=0.4.1
huggingface-hub~=0.33.4
idna~=3.10
itsdangerous~=2.2.0
Jinja2~=3.1.2
jiter~=0.10.0
joblib~=1.5.1
jsonpatch~=1.33
jsonpointer~=3.0.0
kiwisolver~=1.4.8
langchain~=0.3.26
langchain-community~=0.3.26
langchain-core~=0.3.66
langchain-openai~=0.3.24
langchain-anthropic~=0.3.18
langchain-google-genai~=2.1.9
langchain-text-splitters~=0.3.8
langgraph~=0.4.8
langgraph-checkpoint~=2.1.0
langgraph-prebuilt~=0.2.2
langgraph-sdk~=0.1.70
langsmith~=0.4.1
lz4~=4.4.4
MarkupSafe~=3.0.2
marshmallow~=3.26.1
matplotlib~=3.10.3
mpmath~=1.3.0
msal~=1.32.3
msal-extensions~=1.3.1
msgspec~=0.19.0
multidict~=6.4.4
mypy_extensions~=1.1.0
networkx~=3.5
numpy~=1.26.4
oauthlib~=3.3.1
openai~=1.90.0
openpyxl~=3.1.5
orjson~=3.10.18
ormsgpack~=1.10.0
packaging~=24.2
pandas~=2.1.4
patsy~=1.0.1
pillow~=11.2.1
plotly~=5.17.0
propcache~=0.3.2
pyasn1~=0.6.1
pyasn1_modules~=0.4.2
pycparser~=2.22
pydantic~=2.11.7
pydantic-settings~=2.10.1
pydantic_core~=2.33.2
PyJWT~=2.10.1
pyparsing~=3.2.3
python-dateutil~=2.9.0.post0
python-decouple~=3.8
python-dotenv~=1.0.0
python-multipart~=0.0.6
pytz~=2025.2
PyYAML~=6.0.2
regex~=2024.11.6
requests~=2.32.4
requests-toolbelt~=1.0.0
rsa~=4.9.1
safetensors~=0.5.3
scikit-learn~=1.7.0
scipy~=1.15.3
seaborn~=0.13.2
sentence-transformers~=5.0.0
six~=1.17.0
sniffio~=1.3.1
SQLAlchemy~=2.0.41
sqlparse~=0.5.3
statsmodels~=0.14.4
sympy~=1.14.0
tenacity~=9.1.2
threadpoolctl~=3.6.0
thrift~=0.20.0
tiktoken~=0.9.0
tokenizers~=0.21.2
torch~=2.7.1
tqdm~=4.67.1
transformers~=4.53.2
typing-inspect~=0.9.0
typing-inspection~=0.4.1
typing_extensions~=4.14.0
tzdata~=2025.2
urllib3~=2.5.0
uvicorn~=0.24.0
Werkzeug~=3.0.1
xlrd~=2.0.2
xlsxwriter~=3.2.5
xxhash~=3.5.0
yarl~=1.20.1
zstandard~=0.23.0
fuzzywuzzy~=0.18.0
python-Levenshtein~=0.27.1
jellyfish==1.2.0
xgboost~=3.0.4
pyod~=2.0.5
lightgbm~=4.6.0
catboost~=1.2.8
prophet~=1.1.7
shap~=0.48.0
lime~=0.2.0.1
optuna~=4.4.0
pycaret~=3.3.2
category_encoders~=2.7.0
feature-engine~=1.6.2
wordcloud~=1.9.4

